requirements.txt
numpy
Pillow
torch
transformers
timm
safetensors
watchdog
diffusers
